POC: emit a list of PRs
This commit is a proof of concept to enable setting a dynamic list of
instanced pipelines - one per PR. It depends on concourse/concourse#7252
in order to work.

Rather than listing a list of versions (one per PR), we emit a single
version consisting of the list of PR numbers. We can then iterate
`across` this list and `set_pipeline` for each PR. When a PR is removed
from the list, Concourse will automatically archive it based on the
set_pipeline semantics.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>