encode PR data as JSON objects
in the get response file, we previously just wrote the list of PR
numbers to the response file as a JSON list of numbers.

with this change, we encode a JSON list of objects, each with the key
`"number"`

e.g. instead of

[1, 2, 3]

we now do

[{"number": 1}, {"number": 2}, {"number": 3}]

while this may seem like a silly change, it will allow us to add more
metadata to the PR list transparently without breaking users' pipelines.
for instance, if we wanted to include the PR author as well for
iterating with `across`, we can do that:

[{"number": 1, "author": "aoldershaw"}, ...]

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>